Prototype multiple elements on the same line
When a note or signal is indented with '=' it is shown on the same height as
the latest note or signal without indentation.

This is just a PoC implementation of #130. The formatting is mucked up and the
logic is a bit hard to read. If we decide this is a useful feature we probably
want to make 'lines' (collections of notes/signals that are on the same line)
a first-class citizen.