Added support for UML ExecutionSpecifications
I've added support for UML ExecutionSpecifications, which are the grey
bars commonly seen in UML Sequence Diagrams, and indicate the execution
of an action.

Using them is quite simple, all you need to do is prepend a `+` or `-`
character before the actor/lifeline that will be executing an action.
The `+` character will begin the ExecutionSpecification, and the `-`
character will end it.

Example
-------
The following example shows an API executing a function in response to
a user calling function foo().

    User->+API: foo()
    note right of foo: Processing foo.
    -API-->User: Result.